Weapon Class changes, Hit Effects and a Sniper Rifle!
This is a major addition to the Zombie game. Any subclass of Weapon can
now specify if the rays cast from the gun can hit more than one enemy (
new sniper rifle ) and takes a particle effect, or hit effect, for when
an object is shot. An error is thrown if no hitEffect is passed. A
sniper rifle is now in the game and has a pickup