Zombie Level, Vampires and some rebalance!
I decided that just fighting zombies was boring. Why not fight something
a little more scary? Vampires ( these guys are hard )! I also made the
test area a little more visually appealing, so rendering on some
computers may slow down performance. I like the direction that this is
going in. Ultimately, if others are still interested in this, I want
this to turn into a modern witcher. Fighting creatures is always fun.
I'm also working on melee weapons and an ally to assist the player.
Someday i might even make a ui or the player take damage :-P