Buffs and Debuffs
I added a Buff/Debuff system to the zombie game. It works well enough
where it could be added to the main game if others were interested in
using it. Pistol does bonus damage to Vampires and the shotgun applies a
.5 second slow to all enemies hit. Added the beginning of what will be
the usable item slot for the zombie hunter. Also started on the dialogue
system between player and npcs. I will hopefully have most of it if not
all finished by Fridays meeting.